requirements.txt
fastapi==0.108.0
pydantic==2.5.3
firebase_admin==6.5.0
requests==2.31.0
streamlit==1.34.0
st-pages==0.4.5
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dotenv==1.0.1